requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
requests==2.32.3
